model/requirements.txt
openai
sentence-transformers
transformers
python-dotenv